portfolioproject/requirements.txt
absl-py==2.2.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2
dm-tree==0.1.9
exceptiongroup==1.3.0
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.9
nltk==3.9.1
numpy==1.24.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==1.5.3
pillow==11.2.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.45.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.1
